+++ frontend/src/components/home/home.jsx
@@ -1,35 +1,9 @@
-import React from "react";
-import "./home.css";
-import { useNavigate } from 'react-router-dom';
+import React from 'react'
 
-function Home() {
-    const navigate = useNavigate();
-
-    function move(){
-        navigate('log');
-    }
-
+function home() {
   return (
-    <div>
-      <div className="hero min-h-screen bg-base-200">
-        <div className="hero-content flex-col lg:flex-row">
-          <img
-            src="https://placeimg.com/260/400/arch"
-            className="max-w-sm rounded-lg shadow-2xl"
-          />
-          <div>
-            <h1 className="text-5xl font-bold">Movie Theatre</h1>
-            <p className="py-6">
-              Provident cupiditate voluptatem et in. Quaerat fugiat ut assumenda
-              excepturi exercitationem quasi. In deleniti eaque aut repudiandae
-              et a id nisi.
-            </p>
-            <button className="btn btn-primary" onClick={move}>Get Started</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
+    <div>home</div>
+  )
 }
 
-export default Home;
+export default home
