+++ frontend/src/components/splashpage/Splash.jsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { useNavigate } from 'react-router-dom';
+
+function Splash() {
+    const navigate = useNavigate();
+
+    function move(){
+        navigate('log');
+    }
+  return (
+    <div>
+      <div className="hero min-h-screen bg-base-200">
+        <div className="hero-content flex-col lg:flex-row">
+          <img
+            src="https://placeimg.com/260/400/arch"
+            className="max-w-sm rounded-lg shadow-2xl"
+          />
+          <div>
+            <h1 className="text-5xl font-bold">Movie Theatre</h1>
+            <p className="py-6">
+              Provident cupiditate voluptatem et in. Quaerat fugiat ut assumenda
+              excepturi exercitationem quasi. In deleniti eaque aut repudiandae
+              et a id nisi.
+            </p>
+            <button className="btn btn-primary" onClick={move}>
+              Get Started
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Splash;
